Fix non-JavaScript field label generation.
Ensure that when generating the date time component, the nominal field label is generated for the displayed text field. The hidden field which is displayed using JavaScript should be given the hidden ID so that non-JavaScript browsers have the label properly set for the displayed field. Furthermore, the JavaScript needs to reassign the label to the hidden field before displaying the enhanced date-time picker.